Add files via upload
Write a program that receives an integer as the radius of the circle sector and also the angle of the sector (from 1 to 360) and returns the area of ​​the sector. Note: Consider the number pi as three and get the area to two decimal places.

Input 
The input consists of a line that starts with the number
�
rAs the radius and dimension of that number
�
aIt is given as an angle in degrees.
0
≤
�
0≤r
1
≤
�
≤
3
6
0
1≤a≤3 6 0

Output 
The output consists of a line where the sector area is displayed.